AM-004 Add unit test for front end components
This commit adds some unit tests for the front end components
added in AM-003. Note, only a core selection of unit test have
been written for. The components that dont have test require
similar tests to the one that are already written.

Note: this does not include integration test or functional
test. These test check that the components render correctly
when the right props are given.